[FIX] Profile decomposed
